# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Platform - Quant-Service [Development]",
    "description": "Data Service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/market-data/fundamental/overview/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "fundamental"
        ],
        "summary": "기업 개요 조회",
        "description": "지정된 종목의 기업 개요 정보를 조회합니다.",
        "operationId": "Market Data v2-get_company_overview",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Company Overview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/fundamental/income-statement/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "fundamental"
        ],
        "summary": "손익계산서 조회",
        "description": "지정된 종목의 손익계산서를 조회합니다.",
        "operationId": "Market Data v2-get_income_statement",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "보고 주기 (annual, quarterly)",
              "default": "annual",
              "title": "Period"
            },
            "description": "보고 주기 (annual, quarterly)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Income Statement"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/fundamental/balance-sheet/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "fundamental"
        ],
        "summary": "재무상태표 조회",
        "description": "지정된 종목의 재무상태표를 조회합니다.",
        "operationId": "Market Data v2-get_balance_sheet",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "보고 주기 (annual, quarterly)",
              "default": "annual",
              "title": "Period"
            },
            "description": "보고 주기 (annual, quarterly)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Balance Sheet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/fundamental/cash-flow/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "fundamental"
        ],
        "summary": "현금흐름표 조회",
        "description": "지정된 종목의 현금흐름표를 조회합니다.",
        "operationId": "Market Data v2-get_cash_flow",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "보고 주기 (annual, quarterly)",
              "default": "annual",
              "title": "Period"
            },
            "description": "보고 주기 (annual, quarterly)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Cash Flow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/fundamental/earnings/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "fundamental"
        ],
        "summary": "실적 데이터 조회",
        "description": "지정된 종목의 실적 데이터를 조회합니다.",
        "operationId": "Market Data v2-get_earnings",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Earnings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}